+++ dessia_common/core.py
@@ -3,6 +3,7 @@
 """
 
 """
+import builtins
 import sys
 import warnings
 import math
@@ -13,14 +14,22 @@ import collections
 from copy import deepcopy
 import inspect
 import json
-from typing import List, Tuple, Union, Any, \
+import bson
+from openpyxl.writer.excel import save_virtual_workbook
+from openpyxl.styles.borders import Border, Side
+from openpyxl.styles import Alignment, PatternFill
+from openpyxl import Workbook
+import openpyxl.utils
+
+from typing import List, Dict, Type, Tuple, Union, Any, \
     get_type_hints, get_origin, get_args
 try:
     from typing import TypedDict  # >=3.8
 except ImportError:
     from mypy_extensions import TypedDict  # <=3.7
 import traceback as tb
-from dessia_common.typings import Measure, JsonSerializable, Subclass
+from dessia_common.typings import Measure, JsonSerializable,\
+    Subclass, InstanceOf
 
 from importlib import import_module
 
@@ -84,7 +93,7 @@ def deprecated(use_instead=None):
         def wrapper(*args, **kwargs):
             deprecation_warning(function.__name__, 'Function', use_instead)
             print('Traceback : ')
-            tb.print_stack(limit=1)
+            tb.print_stack(limit=2)
             return function(*args, **kwargs)
 
         return wrapper
@@ -157,7 +166,6 @@ class DessiaObject:
     _eq_is_data_eq = True
     
     _init_variables = None
-    _export_formats = None
     _allowed_methods = []
     _whitelist_attributes = []
 
@@ -185,11 +193,11 @@ class DessiaObject:
         if full_classname(self) != full_classname(other_object):
             return False
 
-        eq_dict = self.__getstate__()
+        eq_dict = self._serializable_dict()
         if 'name' in eq_dict:
             del eq_dict['name']
             
-        other_eq_dict = other_object.__getstate__()
+        other_eq_dict = other_object._serializable_dict()
 
         for key, value in eq_dict.items():
             other_value = other_eq_dict[key]
@@ -202,7 +210,7 @@ class DessiaObject:
         forbidden_keys = (self._non_data_eq_attributes
                           + self._non_data_hash_attributes
                           + ['package_version', 'name'])
-        for key, value in self.__getstate__().items():
+        for key, value in self._serializable_dict().items():
             if key not in forbidden_keys:
                 if isinstance(value, list):
                     hash_ += list_hash(value)
@@ -225,9 +233,9 @@ class DessiaObject:
         # eq_dict = {k: v for k, v in self.to_dict().items()
         #            if (k not in ['package_version', 'name'])\
         #                and (k not in self._non_data_eq_attributes)}
-        eq_dict = self.__getstate__()
+        eq_dict = self._serializable_dict()
         # other_eq_dict = other_object.to_dict()
-        other_eq_dict = other_object.__getstate__()
+        other_eq_dict = other_object._serializable_dict()
 
         for key, value in eq_dict.items():
             if key not in other_eq_dict:
@@ -260,17 +268,19 @@ class DessiaObject:
             dict_['package_version'] = package_version
         return dict_
 
-    def __getstate__(self):
+    def _serializable_dict(self):
+
         dict_ = {k: v for k, v in self.__dict__.items()
                  if k not in self._non_serializable_attributes
                  and not k.startswith('_')}
         return dict_
 
+
     def to_dict(self) -> JsonSerializable:
         """
         Generic to_dict method
         """
-        dict_ = self.__getstate__()
+        dict_ = self._serializable_dict()
         if hasattr(self, 'Dict'):
             # !!! This prevent us to call DessiaObject.to_dict()
             # from an inheriting object which implement a Dict method,
@@ -286,7 +296,8 @@ class DessiaObject:
         return serialized_dict
 
     @classmethod
-    def dict_to_object(cls, dict_: JsonSerializable) -> 'DessiaObject':
+    def dict_to_object(cls, dict_: JsonSerializable,
+                       force_generic: bool = False) -> 'DessiaObject':
         """
         Generic dict_to_object method
         """
@@ -296,10 +307,11 @@ class DessiaObject:
             return cls.DictToObject(dict_)
 
         if cls is not DessiaObject:
-            obj = dict_to_object(dict_, cls)
+            obj = dict_to_object(dict_=dict_, class_=cls,
+                                 force_generic=force_generic)
             return obj
         elif 'object_class' in dict_:
-            obj = dict_to_object(dict_)
+            obj = dict_to_object(dict_=dict_, force_generic=force_generic)
             return obj
         else:
             # Using default
@@ -339,12 +351,25 @@ class DessiaObject:
 
         unordered_count = 0
 
+        # Parse docstring
+        try:
+            parsed_docstring = parse_docstring(cls)
+        except Exception:
+            parsed_docstring = {
+                'description': 'Docstring parsing failed',
+                'attributes': {}
+            }
+        parsed_attributes = parsed_docstring['attributes']
+
         # Initialize jsonschema
         _jsonschema = deepcopy(JSONSCHEMA_HEADER)
 
         required_arguments, default_arguments = inspect_arguments(method=init,
                                                                   merge=False)
         _jsonschema['required'] = required_arguments
+        _jsonschema['standalone_in_db'] = cls._standalone_in_db
+        _jsonschema['description'] = parsed_docstring['description']
+        _jsonschema['python_typing'] = str(cls)
 
         # Set jsonschema
         for annotation in annotations.items():
@@ -363,19 +388,22 @@ class DessiaObject:
                 editable = name not in cls._non_editable_attributes
                 annotation_type = type_from_annotation(annotation[1], cls)
                 annotation = (annotation[0], annotation_type)
-                jss_elt = jsonschema_from_annotation(annotation=annotation,
-                                                     jsonschema_element={},
-                                                     order=order,
-                                                     editable=editable,
-                                                     title=title)
+                jss_elt = jsonschema_from_annotation(
+                    annotation=annotation, jsonschema_element={},
+                    order=order, editable=editable, title=title
+                )
+                if name in parsed_attributes:
+                    description = parsed_attributes[name]['desc']
+                    jss_elt[name]['description'] = description
                 _jsonschema['properties'].update(jss_elt)
                 if name in default_arguments.keys():
                     default = set_default_value(_jsonschema['properties'],
                                                 name,
                                                 default_arguments[name])
                     _jsonschema['properties'].update(default)
-                _jsonschema['classes'] = [cls.__module__ + '.' + cls.__name__]
-                _jsonschema['whitelist_attributes'] = cls._whitelist_attributes
+
+        _jsonschema['classes'] = [cls.__module__ + '.' + cls.__name__]
+        _jsonschema['whitelist_attributes'] = cls._whitelist_attributes
         return _jsonschema
 
     @property
@@ -406,6 +434,7 @@ class DessiaObject:
                 if annotations:
                     jsonschemas[method_name] = deepcopy(JSONSCHEMA_HEADER)
                     jsonschemas[method_name]['required'] = []
+                    jsonschemas[method_name]['method'] = True
                     for i, annotation in enumerate(annotations.items()):
                         # TOCHECK Not actually ordered
                         argname = annotation[0]
@@ -459,9 +488,17 @@ class DessiaObject:
         return arguments
 
     def save_to_file(self, filepath, indent=0):
-        with open(filepath + '.json', 'w') as file:
-            json.dump(self.to_dict(), file, indent=indent)
-
+        if isinstance(filepath, str):
+            if not filepath.endswith('.json'):
+                filepath += '.json'
+            file = open(filepath, 'w')
+        else:
+            file = filepath
+        json.dump(self.to_dict(), file, indent=indent)
+        
+        if isinstance(filepath, str):
+            file.close()    
+        
     @classmethod
     def load_from_file(cls, filepath):
         if isinstance(filepath, str):
@@ -509,62 +546,47 @@ class DessiaObject:
                 dict_[arg] = deepcopy_value(getattr(self, arg), memo=memo)
         return self.__class__(**dict_)
 
-    def volmdlr_volume_model(self, frame=None):
+    def volmdlr_volume_model(self, **kwargs):
         if hasattr(self, 'volmdlr_primitives'):
             import volmdlr as vm  # !!! Avoid circular imports, is this OK ?
             if hasattr(self, 'volmdlr_primitives_step_frames'):
-                return vm.MovingVolumeModel(self.volmdlr_primitives(),
-                                            self.volmdlr_primitives_step_frames())
-            else:
-                if frame is None:
-                    frame = vm.OXYZ
-            try:
-                return vm.core.VolumeModel(self.volmdlr_primitives(frame=frame))
-            except TypeError:
-                return vm.core.VolumeModel(self.volmdlr_primitives())
+                return vm.core.MovingVolumeModel(self.volmdlr_primitives(**kwargs),
+                                            self.volmdlr_primitives_step_frames(**kwargs))
+            # else:
+            #     if frame is None:
+            #         frame = vm.OXYZ
+            # try:
+            return vm.core.VolumeModel(self.volmdlr_primitives(**kwargs))
+            # except TypeError:
+            #     return vm.core.VolumeModel(self.volmdlr_primitives())
         msg = 'Object of type {} does not implement volmdlr_primitives'
         raise NotImplementedError(msg.format(self.__class__.__name__))
 
-    def cad_export(self, fcstd_filepath=None, istep=0, python_path='python3',
-                   freecad_lib_path='/usr/lib/freecad/lib', export_types=None):
-        """
-        Generic CAD export method
-        """
-        if fcstd_filepath is None:
-            fcstd_filepath = 'An unnamed {}'.format(self.__class__.__name__)
 
-        if export_types is None:
-            export_types = ['fcstd']
-
-        if hasattr(self, 'volmdlr_primitives'):
-            model = self.volmdlr_volume_model()
-            if model.__class__.__name__ == 'MovingVolumeModel':
-                model = model.step_volume_model(istep)
-            model.freecad_export(fcstd_filepath, python_path=python_path,
-                                 freecad_lib_path=freecad_lib_path,
-                                 export_types=export_types)
-        else:
-            raise NotImplementedError
-
-    def plot(self):
+    def plot(self, **kwargs):
         """
 
         """
         if hasattr(self, 'plot_data'):
             import plot_data
-            for data in self.plot_data():
+            for data in self.plot_data(**kwargs):
                 plot_data.plot_canvas(plot_data_object=data,
                                       canvas_id='canvas',
+                                      width=1400, height=900,
                                       debug_mode=False)
         else:
             msg = 'Class {} does not implement a plot_data method' \
                   'to define what to plot'
             raise NotImplementedError(msg.format(self.__class__.__name__))
 
-    def mpl_plot(self):
+    def mpl_plot(self, **kwargs):
         axs = []
         if hasattr(self, 'plot_data'):
-            for data in self.plot_data():
+            try:
+                plot_datas = self.plot_data(**kwargs)
+            except TypeError as error:
+                raise TypeError('{}.{}'.format(self.__class__.__name__, error))
+            for data in plot_datas:
                 if hasattr(data, 'mpl_plot'):
                     ax = data.mpl_plot()
                     axs.append(ax)
@@ -575,8 +597,8 @@ class DessiaObject:
 
         return axs
 
-    def babylonjs(self, use_cdn=True, debug=False):
-        self.volmdlr_volume_model().babylonjs(use_cdn=use_cdn, debug=debug)
+    def babylonjs(self, use_cdn=True, debug=False, **kwargs):
+        self.volmdlr_volume_model(**kwargs).babylonjs(use_cdn=use_cdn, debug=debug)
 
     def _displays(self, **kwargs) -> List[JsonSerializable]:
         if hasattr(self, '_display_angular'):
@@ -608,11 +630,118 @@ class DessiaObject:
                     display_ = DisplayObject(type_='plot_data', data=plot,
                                              reference_path=reference_path)
                     displays.append(display_.to_dict())
+            else:
+                msg = 'plot_data must return a sequence. Found {}'
+                raise ValueError(msg.format(type(plot_data)))
         if hasattr(self, 'to_markdown'):
             display_ = DisplayObject(type_='markdown', data=self.to_markdown(),
                                      reference_path=reference_path)
             displays.append(display_.to_dict())
         return displays
+
+    def _check_platform(self):
+        """
+        Reproduce lifecycle on platform (serialization, display)
+        """
+        self.dict_to_object(json.loads(json.dumps(self.to_dict())))
+        bson.BSON.encode(self.to_dict())
+        json.dumps(self._displays())
+
+    def to_xlsx(self, filepath):
+        max_column_width = 40
+        color_dessIA1 = "95B3D8"
+        color_dessIA2 = "78909C"
+        grey1 = "CCCCCC"
+        thin_border = Border(left=Side(style='thin'),
+                             right=Side(style='thin'),
+                             top=Side(style='thin'),
+                             bottom=Side(style='thin'))
+
+        pattern_color1 = PatternFill(
+            fill_type="solid",
+            start_color=color_dessIA1,
+            end_color=color_dessIA1)
+
+        pattern_color2 = PatternFill(
+            fill_type="solid",
+            start_color=color_dessIA2,
+            end_color=color_dessIA2)
+
+
+
+        wb = Workbook()
+        # grab the active worksheet
+        ws1 = wb.active
+        ws1.title = 'Object {}'.format(self.__class__.__name__)
+
+        ws1['A1'] = 'Module'
+        ws1['B1'] = 'Class'
+        ws1['C1'] = 'name'
+
+        ws1['A1'].border = thin_border
+        ws1['B1'].border = thin_border
+        ws1['C1'].border = thin_border
+        ws1['A1'].fill = pattern_color1
+        ws1['B1'].fill = pattern_color1
+        ws1['C1'].fill = pattern_color1
+
+
+        ws1['A2'] = self.__module__
+        ws1['B2'] = self.__class__.__name__
+        ws1['C2'] = self.name
+
+        ws1['A2'].border = thin_border
+        ws1['B2'].border = thin_border
+        ws1['C2'].border = thin_border
+        ws1['A2'].fill = pattern_color1
+        ws1['B2'].fill = pattern_color1
+        ws1['C2'].fill = pattern_color1
+
+
+
+        ws1['A4'] = 'Attribute'
+        ws1['A5'] = 'Value'
+        ws1['A4'].border = thin_border
+        ws1['A5'].border = thin_border
+
+
+        # name_column_width = 0
+        i = 1
+        for (k, v) in sorted(self.__dict__.items()):
+            if (not k.startswith('_')) and k != 'name':
+                cell1 = ws1.cell(row=4, column=i, value=k)
+                cell2 = ws1.cell(row=5, column=i, value=str(v))
+
+                cell1.border = thin_border
+                cell1.fill = pattern_color2
+                cell2.border = thin_border
+
+                i += 1
+
+                column_width = min((len(k) + 1.5), max_column_width)
+                column_name = openpyxl.utils.cell.get_column_letter(i)
+                ws1.column_dimensions[column_name].width = column_width
+
+        wb.save("{}.xlsx".format(filepath))
+
+    def to_step(self, filepath):
+        """
+        filepath can be a str or an io.StringIO
+        """
+        return self.volmdlr_volume_model().to_step(filepath=filepath)
+
+    def _export_formats(self):
+        formats = [('json', 'save_to_file'),
+                   ('xlsx', 'to_xlsx')]
+        if hasattr(self, 'volmdlr_primitives'):
+            formats.append(('step', 'to_step'))
+        return formats
+
+
+class Catalog(DessiaObject):
+    def __init__(self, objects: List[DessiaObject], name: str = ''):
+        self.objects = objects
+        DessiaObject.__init__(self, name=name)
 
 
 class DisplayObject(DessiaObject):
@@ -926,17 +1055,18 @@ def get_python_class_from_class_name(full_class_name):
     return class_
 
 
-def dict_to_object(dict_, class_=None):
+def dict_to_object(dict_, class_=None, force_generic: bool = False):
     working_dict = dict_.copy()
     if class_ is None and 'object_class' in working_dict:
         class_ = get_python_class_from_class_name(working_dict['object_class'])
 
-    if class_ is not None:
-        if hasattr(class_, 'dict_to_object') \
-                and (class_.dict_to_object.__func__
-                     is not DessiaObject.dict_to_object.__func__):
+    if class_ is not None and issubclass(class_, DessiaObject):
+        different_methods = (class_.dict_to_object.__func__
+                             is not DessiaObject.dict_to_object.__func__)
+        if different_methods and not force_generic:
             obj = class_.dict_to_object(dict_)
             return obj
+
         if class_._init_variables is None:
             class_argspec = inspect.getfullargspec(class_)
             init_dict = {k: v for k, v in working_dict.items()
@@ -1088,17 +1218,32 @@ def serialize_typing(typing_):
     if is_typing(typing_):
         origin = get_origin(typing_)
         args = get_args(typing_)
-        if origin is list:
+        if origin is Union:
+            if len(args) == 2 and type(None) in args:
+                # This is a false Union => Is a default value set to None
+                return serialize_typing(args[0])
+            else:
+                # Types union
+                argnames = ', '.join([type_fullname(a) for a in args])
+                return 'Union[{}]'.format(argnames)
+        elif origin is list:
             return 'List[' + type_fullname(args[0]) + ']'
         elif origin is tuple:
             argnames = ', '.join([type_fullname(a) for a in args])
             return 'Tuple[{}]'.format(argnames)
+        elif origin is dict:
+            key_type = type_fullname(args[0])
+            value_type = type_fullname(args[1])
+            return 'Dict[{}, {}]'.format(key_type, value_type)
+        elif origin is InstanceOf:
+            return 'InstanceOf[{}]'.format(type_fullname(args[0]))
         else:
             msg = 'Serialization of typing {} is not implemented'
             raise NotImplementedError(msg.format(typing_))
     if isinstance(typing_, type):
         return full_classname(typing_, compute_for='class')
-    raise NotImplementedError('{} of type {}'.format(typing_, type(typing_)))
+    return str(typing_)
+    # raise NotImplementedError('{} of type {}'.format(typing_, type(typing_)))
 
 
 def type_from_argname(argname):
@@ -1138,6 +1283,13 @@ def deserialize_typing(serialized_typing):
                            "workflow non-block variables.")
                     raise TypeError(msg)
             return Tuple[type_from_argname(full_argname)]
+        elif splitted_type[0] == 'Dict':
+            args = full_argname.split(', ')
+            key_type = type_from_argname(args[0])
+            value_type = type_from_argname(args[1])
+            return Dict[key_type, value_type]
+        # elif splitted_type[0] == 'Union':
+        #     args = full_argname.split(', ')
     raise NotImplementedError('{}'.format(serialized_typing))
 
 
@@ -1283,6 +1435,10 @@ def is_sequence(obj):
         and not isinstance(obj, str)
 
 
+def is_builtin(type_):
+    return type_ in TYPING_EQUIVALENCES
+
+
 def type_from_annotation(type_, module):
     """
     Clean up a proposed type if there are stringified
@@ -1298,8 +1454,11 @@ def type_from_annotation(type_, module):
 
 
 def is_typing(object_: Any):
-    in_typings = object_.__module__ in ['typing', 'dessia_common.typings']
     has_module = hasattr(object_, '__module__')
+    if has_module:
+        in_typings = object_.__module__ in ['typing', 'dessia_common.typings']
+    else:
+        return False
     has_origin = hasattr(object_, '__origin__')
     has_args = hasattr(object_, '__args__')
     return has_module and has_origin and has_args and in_typings
@@ -1316,11 +1475,15 @@ def jsonschema_from_annotation(annotation, jsonschema_element,
     if editable is None:
         editable = key not in ['return']
 
+    # Compute base entries
+    jsonschema_element[key] = {'title': title, 'editable': editable,
+                               'order': order,
+                               'python_typing': serialize_typing(typing_)}
+
     if typing_ in TYPING_EQUIVALENCES.keys():
         # Python Built-in type
-        jsonschema_element[key] = {'type': TYPING_EQUIVALENCES[typing_],
-                                   'title': title, 'editable': editable,
-                                   'order': order}
+        jsonschema_element[key]['type'] = TYPING_EQUIVALENCES[typing_]
+
     elif is_typing(typing_):
         origin = get_origin(typing_)
         args = get_args(typing_)
@@ -1336,49 +1499,76 @@ def jsonschema_from_annotation(annotation, jsonschema_element,
                 # Types union
                 classnames = [full_classname(object_=a, compute_for='class')
                               for a in args]
-                jsonschema_element[key] = {
-                    'type': 'object', 'title': title, 'classes': classnames,
-                    'editable': editable, 'order': order
-                }
+
+                standalone_args = [a._standalone_in_db for a in args]
+                if all(standalone_args):
+                    standalone = True
+                elif not any(standalone_args):
+                    standalone = False
+                else:
+                    msg = "standalone_in_db values for type '{}'" \
+                          " are not consistent"
+                    raise ValueError(msg.format(typing_))
+                jsonschema_element[key].update({
+                    'type': 'object', 'classes': classnames,
+                    'standalone_in_db': standalone
+                })
         elif origin is list:
             # Homogenous sequences
-            jsonschema_element[key] = jsonschema_sequence_recursion(
+            jsonschema_element[key].update(jsonschema_sequence_recursion(
                 value=typing_, order=order, title=title, editable=editable
-            )
+            ))
         elif origin is tuple:
             # Heterogenous sequences (tuples)
             items = []
             for type_ in args:
                 items.append({'type': TYPING_EQUIVALENCES[type_]})
-            jsonschema_element[key] = {'additionalItems': False,
-                                       'type': 'array', 'items': items,
-                                       'title': title, 'editable': editable,
-                                       'order': order}
+            jsonschema_element[key].update({'additionalItems': False,
+                                            'type': 'array', 'items': items})
         elif origin is dict:
             # Dynamially created dict structure
             key_type, value_type = args
             if key_type != str:
                 # !!! Should we support other types ? Numeric ?
                 raise NotImplementedError('Non strings keys not supported')
-            jsonschema_element[key] = {
-                'type': 'object', 'order': order,
-                'editable': editable, 'title': title,
+            jsonschema_element[key].update({
+                'type': 'object',
                 'patternProperties': {
                     '.*': {
                         'type': TYPING_EQUIVALENCES[value_type]
                     }
                 }
-            }
+            })
         elif origin is Subclass:
+            warnings.simplefilter('once', DeprecationWarning)
+            msg = "\n\nTyping of attribute '{0}' from class {1} "\
+                  "uses Subclass which is deprecated."\
+                  "\n\nUse 'InstanceOf[{2}]' instead of 'Subclass[{2}]'.\n"
+            arg = args[0].__name__
+            warnings.warn(msg.format(key, args[0], arg), DeprecationWarning)
             # Several possible classes that are subclass of another one
             class_ = args[0]
             classname = full_classname(object_=class_, compute_for='class')
-            jsonschema_element[key] = {'type': 'object', 'order': order,
-                                       'subclass_of': classname,
-                                       'title': title, 'editable': editable}
+            jsonschema_element[key].update({
+                'type': 'object', 'instance_of': classname,
+                'standalone_in_db': class_._standalone_in_db
+            })
+        elif origin is InstanceOf:
+            # Several possible classes that are subclass of another one
+            class_ = args[0]
+            classname = full_classname(object_=class_, compute_for='class')
+            jsonschema_element[key].update({
+                'type': 'object', 'instance_of': classname,
+                'standalone_in_db': class_._standalone_in_db
+            })
         else:
             msg = "Jsonschema computation of typing {} is not implemented"
             raise NotImplementedError(msg.format(typing_))
+    elif hasattr(typing_, '__origin__') and typing_.__origin__ is type:
+        jsonschema_element[key].update({
+            'type': 'object', 'is_class': True,
+            'properties': {'name': {'type': 'string'}}
+        })
     elif issubclass(typing_, Measure):
         ann = (key, float)
         jsonschema_element = jsonschema_from_annotation(
@@ -1390,13 +1580,14 @@ def jsonschema_from_annotation(annotation, jsonschema_element,
         classname = full_classname(object_=typing_, compute_for='class')
         if issubclass(typing_, DessiaObject):
             # Dessia custom classes
-            jsonschema_element[key] = {'type': 'object'}
+            jsonschema_element[key].update({
+                'type': 'object',
+                'standalone_in_db': typing_._standalone_in_db
+            })
         else:
             # Statically created dict structure
-            jsonschema_element[key] = static_dict_jsonschema(typing_)
-        jsonschema_element[key].update({'title': title, 'order': order,
-                                        'editable': editable,
-                                        'classes': [classname]})
+            jsonschema_element[key].update(static_dict_jsonschema(typing_))
+        jsonschema_element[key]['classes'] = [classname]
     return jsonschema_element
 
 
@@ -1405,7 +1596,7 @@ def jsonschema_sequence_recursion(value, order: int, title: str = None,
     if title is None:
         title = 'Items'
     jsonschema_element = {'type': 'array', 'order': order,
-                          'editable': editable, 'title': title}
+                          'python_typing': serialize_typing(value)}
 
     items_type = get_args(value)[0]
     if is_typing(items_type) and get_origin(items_type) is list:
@@ -1436,6 +1627,13 @@ def prettyname(namestr):
 
 
 def static_dict_jsonschema(typed_dict, title=None):
+    warnings.simplefilter('once', DeprecationWarning)
+    msg = "\n\nStatic Dict typing is not fully supported.\n" \
+          "This will most likely lead to non predictable behavior" \
+          " or malfunctionning features. \n" \
+          "Define a custom non-standalone class for type '{}'\n\n"
+    classname = full_classname(typed_dict, compute_for='class')
+    warnings.warn(msg.format(classname), DeprecationWarning)
     jsonschema_element = deepcopy(JSONSCHEMA_HEADER)
     jss_properties = jsonschema_element['properties']
 
@@ -1445,6 +1643,12 @@ def static_dict_jsonschema(typed_dict, title=None):
 
     # TOCHECK : Not actually ordered !
     for i, annotation in enumerate(annotations.items()):
+        attribute, typing_ = annotation
+        if not is_builtin(typing_):
+            msg = "Complex structure as Static dict values is not supported." \
+                  "\n Attribute {} got type {} instead of builtin." \
+                  "\n Consider creating a custom class for complex structures."
+            raise TypeError(msg.format(attribute, typing_))
         jss = jsonschema_from_annotation(annotation=annotation,
                                          jsonschema_element=jss_properties,
                                          order=i, title=title)
@@ -1453,18 +1657,40 @@ def static_dict_jsonschema(typed_dict, title=None):
 
 
 def set_default_value(jsonschema_element, key, default_value):
-    if isinstance(default_value, tuple(TYPING_EQUIVALENCES.keys())) \
-            or default_value is None:
+    datatype = datatype_from_jsonschema(jsonschema_element[key])
+    if default_value is None\
+            or datatype in ['builtin', 'heterogeneous_sequence',
+                            'static_dict', 'dynamic_dict']:
         jsonschema_element[key]['default_value'] = default_value
-    elif is_sequence(default_value):
-        # TODO : Tuple should be considered OK for default_value
+    # elif datatype == 'builtin':
+    #     jsonschema_element[key]['default_value'] = default_value
+    # elif datatype == 'heterogeneous_sequence':
+    #     jsonschema_element[key]['default_value'] = default_value
+    elif datatype == 'homogeneous_sequence':
         msg = 'Object {} of type {} is not supported as default value'
         type_ = type(default_value)
         raise NotImplementedError(msg.format(default_value, type_))
-    else:
+    elif datatype in ['standalone_object', 'embedded_object',
+                      'instance_of', 'union']:
         object_dict = default_value.to_dict()
         jsonschema_element[key]['default_value'] = object_dict
     return jsonschema_element
+    # if isinstance(default_value, tuple(TYPING_EQUIVALENCES.keys())) \
+    #         or default_value is None:
+    #     jsonschema_element[key]['default_value'] = default_value
+    # elif is_sequence(default_value):
+    #     if datatype == 'heterogeneous_sequence':
+    #         jsonschema_element[key]['default_value'] = default_value
+    #     else:
+    #         msg = 'Object {} of type {} is not supported as default value'
+    #         type_ = type(default_value)
+    #         raise NotImplementedError(msg.format(default_value, type_))
+    # else:
+        # if datatype in ['standalone_object', 'embedded_object',
+        #                 'subclass', 'union']:
+        # object_dict = default_value.to_dict()
+        # jsonschema_element[key]['default_value'] = object_dict
+        # else:
 
 
 def inspect_arguments(method, merge=False):
@@ -1591,65 +1817,107 @@ def recursive_instantiation(type_, value):
         raise NotImplementedError(type_)
 
 
-def choose_default(jsonschema):
+def default_sequence(array_jsonschema):
+    if is_sequence(array_jsonschema['items']):
+        # Tuple jsonschema
+        if 'default_value' in array_jsonschema:
+            return array_jsonschema['default_value']
+        return [default_dict(v) for v in array_jsonschema['items']]
+    return None
+
+
+def datatype_from_jsonschema(jsonschema):
     if jsonschema['type'] == 'object':
-        default_value = default_dict(jsonschema)
+        if 'classes' in jsonschema:
+            if len(jsonschema['classes']) > 1:
+                return 'union'
+            if 'standalone_in_db' in jsonschema:
+                if jsonschema['standalone_in_db']:
+                    return 'standalone_object'
+                return 'embedded_object'
+            return 'static_dict'
+        if 'instance_of' in jsonschema:
+            return 'instance_of'
+        if 'patternProperties' in jsonschema:
+            return 'dynamic_dict'
+        if 'method' in jsonschema and jsonschema['method']:
+            return 'embedded_object'
+        if 'is_class' in jsonschema and jsonschema['is_class']:
+            return 'class'
+
     elif jsonschema['type'] == 'array':
-        default_value = default_sequence(jsonschema)
-    elif jsonschema['type'] == 'string':
-        default_value = ''
+        if 'additionalItems' in jsonschema\
+                and not jsonschema['additionalItems']:
+            return 'heterogeneous_sequence'
+        return 'homogeneous_sequence'
+
+    elif jsonschema['type'] in ['number', 'string', 'boolean']:
+        return 'builtin'
+    return None
+
+
+def chose_default(jsonschema):
+    datatype = datatype_from_jsonschema(jsonschema)
+    if datatype in ['heterogeneous_sequence', 'homogeneous_sequence']:
+        return default_sequence(jsonschema)
+    elif datatype == 'static_dict':
+        return default_dict(jsonschema)
+    elif datatype in ['standalone_object', 'embedded_object',
+                      'instance_of', 'union']:
+        if 'default_value' in jsonschema:
+            return jsonschema['default_value']
+        return None
     else:
-        default_value = None
-    return default_value
+        return None
 
 
 def default_dict(jsonschema):
     dict_ = {}
-    if 'properties' in jsonschema:
-        for property_, value in jsonschema['properties'].items():
-            if property_ in jsonschema['required']:
-                dict_[property_] = choose_default(value)
+    datatype = datatype_from_jsonschema(jsonschema)
+    if datatype in ['standalone_object', 'embedded_object', 'static_dict']:
+        for property_, jss in jsonschema['properties'].items():
+            if 'default_value' in jss:
+                dict_[property_] = jss['default_value']
             else:
-                default_value = value['default_value']
-                if value['type'] == 'array':
-                    dict_[property_] = []
-                elif value['type'] == 'object' and default_value is None:
-                    dict_[property_] = {}
-                else:
-                    dict_[property_] = default_value
+                dict_[property_] = chose_default(jss)
+    else:
+        return None
     return dict_
 
 
-def default_sequence(array_jsonschema):
-    if 'minItems' in array_jsonschema and 'maxItems' in array_jsonschema \
-            and array_jsonschema['minItems'] == array_jsonschema['maxItems']:
-        number = array_jsonschema['minItems']
-    elif 'minItems' in array_jsonschema:
-        number = array_jsonschema['minItems']
-    elif 'maxItems' in array_jsonschema:
-        number = array_jsonschema['maxItems']
-    else:
-        number = 1
+class ParsedAttribute(TypedDict):
+    desc: str
+    type_: str
+    annotation: str
 
-    if type(array_jsonschema['items']) == list:
-        # Tuple jsonschema
-        return [default_dict(v) for v in array_jsonschema['items']]
 
-    elif array_jsonschema['items']['type'] == 'object':
-        if 'classes' in array_jsonschema['items']:
-            # TOCHECK classes[0]
-            classname = array_jsonschema['items']['classes'][0]
-            class_ = get_python_class_from_class_name(classname)
-            if issubclass(class_, DessiaObject):
-                if class_._standalone_in_db:  # Standalone object
-                    return []
-                # Embedded object
-                default_subdict = default_dict(class_.jsonschema())
-                return [default_subdict] * number
-            # Static Dict
-            dict_jsonschema = static_dict_jsonschema(class_)
-            default_subdict = default_dict(dict_jsonschema)
-            return [default_subdict] * number
-        # Subclasses
-        return [choose_default(array_jsonschema['items'])] * number
-    return [choose_default(array_jsonschema['items'])] * number
+class ParsedDocstring(TypedDict):
+    description: str
+    attributes: Dict[str, ParsedAttribute]
+
+
+def parse_docstring(cls: Type) -> ParsedDocstring:
+    """
+    Parse docstring of given class. Refer to docs to see how docstrings
+    should be built.
+    """
+    annotations = get_type_hints(cls.__init__)
+    docstring = cls.__doc__
+    if docstring:
+        splitted_docstring = docstring.split(':param ')
+        parsed_docstring = {"description": splitted_docstring[0].strip()}
+        params = splitted_docstring[1:]
+        args = {}
+        for param in params:
+            splitted_param = param.split(':type ')
+            arg = splitted_param[0]
+            typestr = splitted_param[1]
+            argname, argdesc = arg.split(":", maxsplit=1)
+            argtype = typestr.split(argname+":")[-1]
+            annotation = annotations[argname]
+            args[argname] = {'desc': argdesc.strip(), 'type_': argtype.strip(),
+                             'annotation': str(annotation)}
+            # TODO Should be serialize typing ?
+        parsed_docstring.update({'attributes': args})
+        return parsed_docstring
+    return {'description': "", 'attributes': {}}