+++ dessia_common/forms.py
@@ -43,7 +43,7 @@ except:
     pass
 
 from dessia_common import DessiaObject
-from dessia_common.typings import Subclass
+from dessia_common.typings import InstanceOf, Distance
 from dessia_common.vectored_objects import Catalog
 
 try:
@@ -51,19 +51,22 @@ try:
 except ImportError:
     from mypy_extensions import TypedDict  # <=3.7
 
+from numpy import linspace
+from math import cos
+
 
 class StandaloneSubobject(DessiaObject):
     _standalone_in_db = True
     _generic_eq = True
 
-    def __init__(self, floatarg: float, name: str = 'Standalone Subobject'):
+    def __init__(self, floatarg: Distance, name: str = 'Standalone Subobject'):
         self.floatarg = floatarg
 
         DessiaObject.__init__(self, name=name)
 
     @classmethod
     def generate(cls, seed: int) -> 'StandaloneSubobject':
-        floatarg = 1.7 * seed
+        floatarg = Distance(1.7 * seed)
         name = 'StandaloneSubobject' + str(seed)
         return cls(floatarg=floatarg, name=name)
 
@@ -86,8 +89,11 @@ class StandaloneSubobject(DessiaObject):
         return volumes
 
 
+DEF_SS = StandaloneSubobject.generate(1)
+
+
 class EnhancedStandaloneSubobject(StandaloneSubobject):
-    def __init__(self, floatarg: float, boolarg: bool,
+    def __init__(self, floatarg: Distance, boolarg: bool,
                  name: str = 'Standalone Subobject'):
         self.boolarg = boolarg
 
@@ -95,14 +101,17 @@ class EnhancedStandaloneSubobject(StandaloneSubobject):
 
     @classmethod
     def generate(cls, seed: int) -> 'EnhancedStandaloneSubobject':
-        floatarg = 1.2 * seed
+        floatarg = Distance(1.2 * seed)
         boolarg = floatarg.is_integer()
         name = 'EnhancedStandaloneSubobject' + str(seed)
         return cls(floatarg=floatarg, boolarg=boolarg, name=name)
 
 
+DEF_ESS = EnhancedStandaloneSubobject.generate(1)
+
+
 class InheritingStandaloneSubobject(StandaloneSubobject):
-    def __init__(self, floatarg: float, strarg: str,
+    def __init__(self, floatarg: Distance, strarg: str,
                  name: str = 'Inheriting Standalone Subobject'):
         self.strarg = strarg
 
@@ -110,10 +119,13 @@ class InheritingStandaloneSubobject(StandaloneSubobject):
 
     @classmethod
     def generate(cls, seed: int) -> 'InheritingStandaloneSubobject':
-        floatarg = 0.7 * seed
+        floatarg = Distance(0.7 * seed)
         strarg = str(-seed)
         name = 'Inheriting Standalone Subobject' + str(seed)
         return cls(floatarg=floatarg, strarg=strarg, name=name)
+
+
+DEF_ISS = InheritingStandaloneSubobject.generate(1)
 
 
 class EmbeddedSubobject(DessiaObject):
@@ -140,16 +152,44 @@ class EmbeddedSubobject(DessiaObject):
         return [cls.generate(i) for i in range(ceil(seed/3))]
 
 
-class StaticDict(TypedDict):
-    name: str
-    value: float
-    is_valid: bool
-    subobject: EmbeddedSubobject
+DEF_ES = EmbeddedSubobject.generate(10)
+
+
+class StaticDict(DessiaObject):
+    def __init__(self, float_value: float, int_value: int,
+                 is_valid: bool, name: str = ''):
+        self.float_value = float_value
+        self.int_value = int_value
+        self.is_valid = is_valid
+        DessiaObject.__init__(self, name=name)
+
+
+DEF_SD = StaticDict(name="Default SD Name", float_value=1.3,
+                    int_value=0, is_valid=True)
+
+
+UnionArg = Union[StandaloneSubobject, EnhancedStandaloneSubobject]
 
 
 class StandaloneObject(DessiaObject):
+    """
+    Dev Object for testing purpose
+
+    :param standalone_subobject: A dev subobject that is standalone_in_db
+    :type standalone_subobject: StandaloneSubobject
+    :param embedded_subobject: A dev subobject that isn't standalone_in_db
+    :type embedded_subobject: EmbeddedSubobject
+    :param dynamic_dict: A variable length dict
+    :type dynamic_dict: Dict[str, bool]
+    :param static_dict: A 1-level structurewith only builtin values & str keys
+    :type static_dict: StaticDict
+    :param tuple_arg: A heterogeneous sequence
+    :type tuple_arg: tuple
+    """
     _standalone_in_db = True
     _generic_eq = True
+    _allowed_methods = ['add_standalone_object',
+                        'add_embedded_object', 'add_float']
 
     def __init__(self, standalone_subobject: StandaloneSubobject,
                  embedded_subobject: EmbeddedSubobject,
@@ -158,9 +198,8 @@ class StandaloneObject(DessiaObject):
                  object_list: List[StandaloneSubobject],
                  subobject_list: List[EmbeddedSubobject],
                  builtin_list: List[int],
-                 union_arg: Union[StandaloneSubobject,
-                                  EnhancedStandaloneSubobject],
-                 inheritance_list: List[Subclass[StandaloneSubobject]] = None,
+                 union_arg: UnionArg,
+                 subclass_arg: InstanceOf[StandaloneSubobject],
                  name: str = 'Standalone Object Demo'):
         self.union_arg = union_arg
         self.builtin_list = builtin_list
@@ -173,7 +212,7 @@ class StandaloneObject(DessiaObject):
         self.dynamic_dict = dynamic_dict
         self.standalone_subobject = standalone_subobject
         self.embedded_subobject = embedded_subobject
-        self.inheritance_list = inheritance_list
+        self.subclass_arg = subclass_arg
 
         DessiaObject.__init__(self, name=name)
 
@@ -183,9 +222,9 @@ class StandaloneObject(DessiaObject):
         standalone_subobject = StandaloneSubobject.generate(seed)
         embedded_subobject = EmbeddedSubobject.generate(seed)
         dynamic_dict = {'n'+str(i): bool(seed % 2) for i in range(seed)}
-        static_dict = {'name': 'Object'+str(seed), 'value': seed * 1.3,
-                       'is_valid': is_even,
-                       'embedded_subobject': embedded_subobject}
+        static_dict = StaticDict(name='Object'+str(seed),
+                                 float_value=seed * 1.3,
+                                 int_value=seed, is_valid=is_even)
         tuple_arg = ('value', seed * 3)
         intarg = seed
         strarg = str(seed) * floor(seed/3)
@@ -193,15 +232,17 @@ class StandaloneObject(DessiaObject):
         subobject_list = EmbeddedSubobject.generate_many(seed)
         builtin_list = [seed]*seed
         union_arg = EnhancedStandaloneSubobject.generate(seed)
-        inheritance_list = [StandaloneSubobject.generate(-seed),
-                            InheritingStandaloneSubobject.generate(seed)]
+        if is_even:
+            subclass_arg = StandaloneSubobject.generate(-seed)
+        else:
+            subclass_arg = InheritingStandaloneSubobject.generate(seed)
         return cls(standalone_subobject=standalone_subobject,
                    embedded_subobject=embedded_subobject,
                    dynamic_dict=dynamic_dict, static_dict=static_dict,
                    tuple_arg=tuple_arg, intarg=intarg, strarg=strarg,
                    object_list=object_list, subobject_list=subobject_list,
                    builtin_list=builtin_list, union_arg=union_arg,
-                   inheritance_list=inheritance_list)
+                   subclass_arg=subclass_arg)
 
     def add_standalone_object(self, object_: StandaloneSubobject):
         """
@@ -219,7 +260,7 @@ class StandaloneObject(DessiaObject):
         """
         self.subobject_list.append(object_)
 
-    def add_float(self, value) -> StandaloneSubobject:
+    def add_float(self, value: float) -> StandaloneSubobject:
         """
         This methods adds value to its standalone subobject
         floatarg property and returns it.
@@ -266,11 +307,88 @@ class StandaloneObject(DessiaObject):
         multi_plot = plot_data.MultiplePlots(elements=points, plots=objects,
                                              sizes=sizes, coords=coords,
                                              name='Multiple Plot')
-        # return [scatter_plot]
-        # return [parallel_plot]
-        # return [multi_plot]
-        # return [scatter_plot, parallel_plot]
-        return [primitives_group, scatter_plot, parallel_plot, multi_plot]
+
+        attribute_names = ['time', 'electric current']
+        tooltip = plot_data.Tooltip(to_disp_attribute_names=attribute_names)
+        time1 = linspace(0, 20, 20)
+        current1 = [t ** 2 for t in time1]
+        elements1 = []
+        for time, current in zip(time1, current1):
+            elements1.append({'time': time, 'electric current': current})
+
+        # The previous line instantiates a dataset with limited arguments but
+        # several customizations are available
+        point_style = plot_data.PointStyle(color_fill=RED, color_stroke=BLACK)
+        edge_style = plot_data.EdgeStyle(color_stroke=BLUE, dashline=[10, 5])
+
+        custom_dataset = plot_data.Dataset(elements=elements1, name='I = f(t)',
+                                           tooltip=tooltip,
+                                           point_style=point_style,
+                                           edge_style=edge_style)
+
+        # Now let's create another dataset for the purpose of this exercice
+        time2 = linspace(0, 20, 100)
+        current2 = [100 * (1 + cos(t)) for t in time2]
+        elements2 = []
+        for time, current in zip(time2, current2):
+            elements2.append({'time': time, 'electric current': current})
+
+        dataset2 = plot_data.Dataset(elements=elements2, name='I2 = f(t)')
+
+        graph2d = plot_data.Graph2D(graphs=[custom_dataset, dataset2],
+                                    to_disp_attribute_names=attribute_names)
+        return [primitives_group, scatter_plot,
+                parallel_plot, multi_plot, graph2d]
+
+    def maldefined_method(self, arg0, arg1=1, arg2: int = 10, arg3 = 3):
+        """
+        Defining a docstring for testing parsing purpose
+        """
+        nok_string = "This is a bad coding behavior"
+        ok_string = "This could be OK as temporary attr"
+        self.maldefined_attr = nok_string
+        self._ok_attribute = ok_string
+
+        computation = nok_string + 'or' + ok_string
+
+        return computation
+
+
+DEF_SO = StandaloneObject.generate(1)
+
+
+class StandaloneObjectWithDefaultValues(StandaloneObject):
+    def __init__(self, standalone_subobject: StandaloneSubobject = DEF_SS,
+                 embedded_subobject: EmbeddedSubobject = DEF_ES,
+                 dynamic_dict: Dict[str, bool] = None,
+                 static_dict: StaticDict = DEF_SD,
+                 tuple_arg: Tuple[str, int] = ("Default Tuple", 0),
+                 intarg: int = 1, strarg: str = "Default Strarg",
+                 object_list: List[StandaloneSubobject] = None,
+                 subobject_list: List[EmbeddedSubobject] = None,
+                 builtin_list: List[int] = None,
+                 union_arg: UnionArg = DEF_ESS,
+                 subclass_arg: InstanceOf[StandaloneSubobject] = DEF_ISS,
+                 name: str = 'Standalone Object Demo'):
+        if dynamic_dict is None:
+            dynamic_dict = {}
+        if object_list is None:
+            object_list = [DEF_SS]
+        if subobject_list is None:
+            subobject_list = [DEF_ES]
+        if builtin_list is None:
+            builtin_list = [1, 2, 3, 4, 5]
+        StandaloneObject.__init__(
+            self, standalone_subobject=standalone_subobject,
+            embedded_subobject=embedded_subobject, dynamic_dict=dynamic_dict,
+            static_dict=static_dict, tuple_arg=tuple_arg, intarg=intarg,
+            strarg=strarg, object_list=object_list,
+            subobject_list=subobject_list, builtin_list=builtin_list,
+            union_arg=union_arg, subclass_arg=subclass_arg, name=name
+        )
+
+
+DEF_SOWDV = StandaloneObjectWithDefaultValues()
 
 
 class Generator(DessiaObject):